Refactor sync functionality into a helper class
