+++ python/pyxet/pyxet/cli.py
@@ -3,22 +3,18 @@ import subprocess
 import sys
 import threading
 import typing
-from concurrent.futures import ThreadPoolExecutor
-from dateutil import parser
-from datetime import datetime
 
-import boto3
-import botocore
-import fsspec
 import typer
 from tabulate import tabulate
 from typing_extensions import Annotated
 
-import pyxet
 from .file_system import XetFS
 from .rpyxet import rpyxet
+from .sync import SyncCommand
 from .url_parsing import parse_url
+from .util import CHUNK_SIZE, _get_fs_and_path, _root_copy
 from .version import __version__
+import pyxet.util as util
 
 cli = typer.Typer(add_completion=True, short_help="a pyxet command line interface", no_args_is_help=True)
 repo = typer.Typer(add_completion=False, short_help="sub-commands to manage repositories")
@@ -26,442 +22,6 @@ branch = typer.Typer(add_completion=False, short_help="sub-commands to manage br
 
 cli.add_typer(repo, name="repo")
 cli.add_typer(branch, name="branch")
-
-MAX_CONCURRENT_COPIES = threading.Semaphore(32)
-CHUNK_SIZE = 16 * 1024 * 1024
-
-
-def _ltrim_match(s, match):
-    """
-    Trims the string 'match' from the left of the string 's'
-    raising an error if the match does not exist.
-    Ex:
-    ```
-       ltrim_match("a/b/c.txt", "a/b") => "/c.txt"
-    ```
-    Used to compute relative paths.
-    """
-    if len(s) < len(match):
-        raise (ValueError(f"Path {s} not in directory {match}"))
-    if s[:len(match)] != match:
-        raise (ValueError(f"Path {s} not in directory {match}"))
-    return s[len(match):]
-
-
-def _should_load_aws_credentials():
-    """
-    Determines if AWS credentials should be loaded for s3 API by checking if credentials are available
-    :return: boolean, True if credentials are available, False
-    """
-    client = boto3.client("sts")
-    try:
-        client.get_caller_identity()
-    except botocore.exceptions.NoCredentialsError:
-        return False
-    return True
-
-
-def _get_fs_and_path(uri) -> tuple[fsspec.spec.AbstractFileSystem, str]:
-    if uri.find('://') == -1:
-        fs = fsspec.filesystem("file")
-        uri = os.path.abspath(uri)
-        return fs, uri
-    split = uri.split("://")
-    if len(split) != 2:
-        print(f"Invalid URL: {uri}", file=sys.stderr)
-    if split[0] == 'xet':
-        fs = pyxet.XetFS()
-    elif split[0] == 's3':
-        load_aws_creds = _should_load_aws_credentials()
-        fs = fsspec.filesystem('s3', anon=not load_aws_creds)
-        # this is *really* annoying But the s3fs protocol has
-        # protocol as a list ['s3','s3a']
-        fs.protocol = 's3'
-    else:
-        fs = fsspec.filesystem(split[0])
-    return fs, split[1]
-
-
-def _single_file_copy(src_fs, src_path, dest_fs, dest_path,
-                      buffer_size=CHUNK_SIZE, size_hint=None):
-    if dest_path.split('/')[-1] == '.gitattributes':
-        print("Skipping .gitattributes as that is required for Xet Magic")
-        return
-    print(f"Copying {src_path} to {dest_path}...")
-
-    if src_fs.protocol == 'xet' and dest_fs.protocol == 'xet':
-        dest_fs.cp_file(src_path, dest_path)
-        return
-    with MAX_CONCURRENT_COPIES:
-        try:
-            if dest_fs.protocol == "xet":
-                if size_hint is None:
-                    size_hint = src_fs.info(src_path).get('size', None)
-
-                # Heuristic for now -- if the size of the source is larger than 50MB,
-                # then make sure we have any shards for the destination that work.
-                if size_hint is not None and size_hint >= 50000000:
-                    dest_fs.add_deduplication_hints(dest_path)
-
-            with src_fs.open(src_path, "rb") as source_file:
-                with dest_fs.open(dest_path, "wb", auto_mkdir=True) as dest_file:
-                    # Buffered copy in chunks
-                    while True:
-                        chunk = source_file.read(buffer_size)
-                        if not chunk:
-                            break
-                        dest_file.write(chunk)
-        except Exception as e:
-            proto = src_fs.protocol
-            print(f"Failed to copy {proto}://{src_path}: {e}")
-
-
-def _validate_xet_copy(src_fs, src_path, dest_fs, dest_path):
-    """
-    Performs some basic early validation of a xet to avoid issues later.
-    Does not catch all failure conditions, but catches some early enough to
-    avoid doing a lot of unnecessary work, then fail.
-
-    Raises an exception on failure, returns True on success
-    """
-    srcproto = src_fs.protocol
-    destproto = dest_fs.protocol
-    # if src is a xet, there must be a branch to copy from
-    if srcproto == 'xet':
-        src_fs.branch_info(src_path)
-
-    if destproto == 'xet':
-        # check dest branch exists before we try to do any copying
-        # An exception is that if this operation would create a branch
-        if srcproto == 'xet':
-            src_parse = parse_url(src_path, src_fs.domain)
-            dest_parse = parse_url(dest_path, dest_fs.domain)
-            if src_parse.path == '' and dest_parse.path == '':
-                # this is a branch to branch copy
-                return True
-
-        dest_fs.branch_info(dest_path)
-
-
-def _validate_xet_sync(source, destination):
-    """
-    Performs early validation for the source and destination paths of
-    a xet sync. Doesn't catch all failure conditions, but catches many
-    of the easy validations.
-
-    Raises exceptions on failure
-    """
-    src_fs, src_path = _get_fs_and_path(source)
-    src_protocol = src_fs.protocol
-    dest_fs, dest_path = _get_fs_and_path(destination)
-    dest_protocol = dest_fs.protocol
-
-    if dest_protocol != 'xet':
-        raise ValueError(f"Unsupported destination protocol: {dest_protocol}, only xet:// targets are supported")
-    if src_protocol == 'xet':
-        raise ValueError(f"Unsupported source protocol: {src_protocol}, only non-xet sources are supported")
-
-    # check that the destination specifies an existing branch
-    # TODO: we may want to be able to sync remote location to a new branch?
-    dest_fs.branch_info(dest_path)
-
-    # source should be a directory
-    if not _isdir(src_fs, src_path):
-        raise ValueError(f"source: {source} needs to be a directory")
-
-    # wildcards not supported
-    if '*' in src_path or '*' in dest_path:
-        raise ValueError(f"Wildcards not supported in path")
-
-
-def _isdir(fs, path):
-    if fs.protocol == 'xet':
-        return fs.isdir_or_branch(path)
-    else:
-        return fs.isdir(path)
-
-
-def _copy(source, destination, recursive=True, _src_fs=None, _dest_fs=None):
-    src_fs, src_path = _get_fs_and_path(source)
-    dest_fs, dest_path = _get_fs_and_path(destination)
-    if _src_fs is not None:
-        src_fs = _src_fs
-    if _dest_fs is not None:
-        dest_fs = _dest_fs
-    srcproto = src_fs.protocol
-    destproto = dest_fs.protocol
-
-    _validate_xet_copy(src_fs, src_path, dest_fs, dest_path)
-
-    # normalize trailing '/' by just removing them unless the path
-    # is exactly just '/'
-    if src_path != '/':
-        src_path = src_path.rstrip('/')
-    if dest_path != '/':
-        dest_path = dest_path.rstrip('/')
-    src_isdir = _isdir(src_fs, src_path)
-
-    # Handling wildcard cases
-    if '*' in src_path:
-        # validate
-        # we only accept globs of the for blah/blah/blah/[glob]
-        # i.e. the glob is only in the last component
-        # src_root_dir should be blah/blah/blah here
-        src_root_dir = '/'.join(src_path.split('/')[:-1])
-        if '*' in src_root_dir:
-            raise ValueError(f"Invalid glob {source}. Wildcards can only appear in the last position")
-        # The source path contains a wildcard
-        with ThreadPoolExecutor() as executor:
-            futures = []
-            for path, info in src_fs.glob(src_path, detail=True).items():
-                # Copy each matching file
-                if info['type'] == 'directory' and not recursive:
-                    continue
-                relpath = _ltrim_match(path, src_root_dir).lstrip('/')
-                if dest_path == '/':
-                    dest_for_this_path = f"/{relpath}"
-                else:
-                    dest_for_this_path = f"{dest_path}/{relpath}"
-                dest_dir = '/'.join(dest_for_this_path.split('/')[:-1])
-                dest_fs.makedirs(dest_dir, exist_ok=True)
-
-                if info['type'] == 'directory':
-                    futures.append(
-                        executor.submit(
-                            _copy,
-                            f"{src_fs.protocol}://{path}",
-                            f"{dest_fs.protocol}://{dest_for_this_path}",
-                            recursive=True,
-                            _src_fs=src_fs,
-                            _dest_fs=dest_fs,
-                        )
-                    )
-                else:
-                    futures.append(
-                        executor.submit(
-                            _single_file_copy,
-                            src_fs,
-                            f"{path}",
-                            dest_fs,
-                            dest_for_this_path,
-                            size_hint=info.get('size', None)
-                        ))
-
-            for future in futures:
-                future.result()
-        return
-
-    # Handling directories
-    if src_isdir:
-        # Recursively copy
-        # xet cp_file can cp directories
-        if srcproto == 'xet' and destproto == 'xet':
-            print(f"Copying {src_path} to {dest_path}...")
-            dest_fs.cp_file(src_path, dest_path)
-            return
-        with ThreadPoolExecutor() as executor:
-            futures = []
-            for path, info in src_fs.find(src_path, detail=True).items():
-                if info['type'] == 'directory' and not recursive:
-                    continue
-                # Note that path is a full path
-                # we need to relativize to make the destination path
-                relpath = _ltrim_match(path, src_path).lstrip('/')
-                if dest_path == '/':
-                    dest_for_this_path = f"/{relpath}"
-                else:
-                    dest_for_this_path = f"{dest_path}/{relpath}"
-                dest_dir = '/'.join(dest_for_this_path.split('/')[:-1])
-                dest_fs.makedirs(dest_dir, exist_ok=True)
-                # Submitting copy jobs to thread pool
-                futures.append(
-                    executor.submit(
-                        _single_file_copy,
-                        src_fs,
-                        f"{path}",
-                        dest_fs,
-                        dest_for_this_path,
-                        size_hint=info.get('size', None)
-                    ))
-            # Waiting for all copy jobs to complete
-            for future in futures:
-                future.result()
-        return
-
-    _single_file_copy(src_fs, src_path, dest_fs, dest_path)
-
-
-def _root_copy(source, destination, message, recursive=False):
-    dest_fs, dest_path = _get_fs_and_path(destination)
-    destproto_is_xet = dest_fs.protocol == 'xet'
-    dest_isdir = _isdir(dest_fs, dest_path)
-
-    # Our target is an existing directory and src is not a wildcard copy
-    # i.e. we are getting cp src/some/path to dest/some/where
-    # but dest/some/where exists as a directory
-    # So we will need to make the dest dest/some/where/path
-    if dest_isdir and '*' not in source:
-        # split up the final component from source path and add it
-        # to the destination
-        final_source_component = source.split('/')[-1]
-        if not destination.endswith('/'):
-            destination += '/'
-        destination += final_source_component
-
-    if destproto_is_xet:
-        dest_fs.start_transaction(message)
-    _copy(source, destination, recursive)
-    if destproto_is_xet:
-        dest_fs.end_transaction()
-
-
-def _get_last_modified(protocol, info):
-    if protocol == 'xet':
-        mod_time = info['last_modified']  # str
-        mod_time = datetime.strptime(mod_time, "%Y-%m-%dT%H:%M:%S%z").timestamp()
-        # mod_time = datetime.fromisoformat(mod_time).timestamp()
-    elif protocol == 's3':
-        mod_time = info['LastModified']  # datetime
-        mod_time = mod_time.timestamp()
-    elif protocol == 'file':
-        mod_time = info['mtime']  # float
-    else:
-        print(f"WARN: protocol: {protocol} doesn't have a modification time, only comparing size")
-        mod_time = None
-
-    return mod_time
-
-
-def _should_sync(src_fs, src_info, dest_fs, dest_info):
-    if src_info['size'] != dest_info['size']:
-        return True
-
-    src_mtime = _get_last_modified(src_fs.protocol, src_info)
-    dest_mtime = _get_last_modified(dest_fs.protocol, dest_info)
-    return src_mtime is not None and dest_mtime is not None \
-        and src_mtime > dest_mtime
-
-
-def _copy_dir_async(executor, futures, src_fs, src_path, dest_fs, dest_path, dryrun):
-    for path, info in src_fs.find(src_path, detail=True).items():
-        # Note that path is a full path
-        # we need to relativize to make the destination path
-        relpath = _ltrim_match(path, src_path).lstrip('/')
-        if dest_path == '/':
-            dest_for_this_path = f"/{relpath}"
-        else:
-            dest_for_this_path = f"{dest_path}/{relpath}"
-        dest_dir = '/'.join(dest_for_this_path.split('/')[:-1])
-        dest_fs.makedirs(dest_dir, exist_ok=True)
-        # Submitting copy jobs to thread pool
-        print(f'copy: {src_fs.protocol}://{path} to {dest_fs.protocol}://{dest_for_this_path}')
-        if not dryrun:
-            futures.append(
-                executor.submit(
-                    _single_file_copy,
-                    src_fs,
-                    f"{path}",
-                    dest_fs,
-                    dest_for_this_path,
-                    size_hint=info.get('size', None)
-                ))
-
-
-def _sync(source, destination, message, dryrun):
-    print(f"Checking sync {source} -> {destination}")
-    _validate_xet_sync(source, destination)
-    print(f"Starting sync {source} -> {destination}")
-
-    src_fs, src_path = _get_fs_and_path(source)
-    dest_fs, dest_path = _get_fs_and_path(destination)
-
-    # normalize trailing '/' by just removing them unless the path
-    # is exactly just '/'
-    if src_path != '/':
-        src_path = src_path.rstrip('/')
-    if dest_path != '/':
-        dest_path = dest_path.rstrip('/')
-
-    failed_copies = []
-
-    if not dryrun:
-        dest_fs.start_transaction(message)
-    with ThreadPoolExecutor() as executor:
-        futures = []
-        # print(f"listing {src_path}")
-        _sync_dir(executor, futures, failed_copies, src_fs, src_path, dest_fs, dest_path, dryrun)
-
-        # Waiting for all copy jobs to complete
-        for future in futures:
-            try:
-                future.result()
-            except Exception as e:
-                print(f"Error: {e}")
-                failed_copies.append("")
-
-    if not dryrun:
-        dest_fs.end_transaction()
-        print(f"Completed sync with: {len(failed_copies)} failed copies")
-
-
-def _sync_dir(executor, futures, failed_copies, src_fs, src_path, dest_fs, dest_path, dryrun):
-    src_proto = src_fs.protocol
-    dest_proto = dest_fs.protocol
-    for src_info in src_fs.ls(src_path, detail=True):
-        # print(f"found: {src_info} from source")
-        abs_path = src_info['name']
-        rel_path = _ltrim_match(abs_path, src_path).lstrip('/')
-        dest_for_this_path = f"/{rel_path}" if dest_path == '/' \
-            else f"{dest_path}/{rel_path}"
-
-        # print(f"checking destination: {dest_for_this_path}")
-        try:
-            dest_info = dest_fs.info(dest_for_this_path)
-        except FileNotFoundError as e:
-            dest_info = None
-
-        # print(f"found: {dest_info} from destination")
-        if dest_info is not None and src_info['type'] != dest_info['type']:
-            failed_copies.append(rel_path)
-            print(f"Copy failed: {src_path} is a {src_info['type']}, {dest_path} is a {dest_info['type']}")
-            continue
-
-        if dest_info is None:
-            # copy src to dest
-            # dest_dir = '/'.join(dest_for_this_path.split('/')[:-1])
-            # dest_fs.makedirs(dest_dir, exist_ok=True)
-            if src_info['type'] == 'directory':
-                print(f"copy dir: {abs_path} to: {dest_for_this_path}")
-                _copy_dir_async(executor, futures, src_fs, abs_path, dest_fs, dest_for_this_path, dryrun)
-            else:
-                print(f"copy: {src_proto}://{abs_path} to: {dest_proto}://{dest_for_this_path}")
-                if not dryrun:
-                    futures.append(
-                        executor.submit(
-                            _single_file_copy,
-                            src_fs,
-                            f"{abs_path}",
-                            dest_fs,
-                            dest_for_this_path,
-                            size_hint=src_info.get('size', None)
-                        ))
-        elif src_info['type'] == 'directory':
-            # recursively sync src -> dest
-            print(f"sync dir: {abs_path} to: {dest_for_this_path}")
-            _sync_dir(executor, futures, failed_copies, src_fs, abs_path, dest_fs, dest_for_this_path, dryrun)
-        elif _should_sync(src_fs, src_info, dest_fs, dest_info):
-            # copy src to dest
-            print(f"copy: {src_proto}://{abs_path} to: {dest_proto}://{dest_for_this_path}")
-            if not dryrun:
-                futures.append(
-                    executor.submit(
-                        _single_file_copy,
-                        src_fs,
-                        f"{abs_path}",
-                        dest_fs,
-                        dest_for_this_path,
-                        size_hint=src_info.get('size', None)
-                    ))
 
 
 class PyxetCLI:
@@ -563,7 +123,7 @@ class PyxetCLI:
         """copy files and folders"""
         if not message:
             message = f"copy {source} to {target}" if not recursive else f"copy {source} to {target} recursively"
-        MAX_CONCURRENT_COPIES = threading.Semaphore(parallel)
+        util.MAX_CONCURRENT_COPIES = threading.Semaphore(parallel)
         _root_copy(source, target, message, recursive=recursive)
 
     @staticmethod
@@ -578,8 +138,10 @@ class PyxetCLI:
         """Copy changed files from source location to destination"""
         if not message:
             message = f"sync {source} to {target}"
-        MAX_CONCURRENT_COPIES = threading.Semaphore(parallel)
-        _sync(source, target, message, dryrun)
+        util.MAX_CONCURRENT_COPIES = threading.Semaphore(parallel)
+        cmd = SyncCommand(source, target, message, dryrun)
+        cmd.validate()
+        cmd.run()
 
     @staticmethod
     @cli.command()
